Add previous events
Changed UpcomingEvents to Events so it will ask for a type and a max. The type will either be upcoming or previous, so Leos can also see the events that have previously happened. If no type is specified, it will show all events. The max is for a max number of events shown

TODO: Add some sort of React form functionality to the home page for new signups. Finish About and Mission page. Instead of having different pages for each of the global causes underneath the mission subroot, have them all on the Mission page.