Create CNAME
npm run deploy (gh-pages command) was overwriting the build directory each time, so the CNAME file is deleted. The gh-pages website would work at sbleos.github.io but not sbleos.org unless a CNAME file exists. A fix is to include CNAME in the public/ directory of the dev branch.