Fix 'xs' Bootstrap class to achieve a properly responsive website
Notes: Bootstrap classes that are responsive such as 'col-{size}-{number}' work with sm, md, lg, and xl. 'xs' is achieved through the default option, so it is not included as an actual class. For example, there is no 'text-xs-center' class. This is instead achieved with 'text-center', which centers content for all screen sizes 'xs' and greater *unless there is another size specified*. So if a tag has a class of 'text-center text-md-left', it will center content on 'xs' and 'sm' screen sizes and left-align content for 'md', 'lg', and 'xl' screen sizes.

TODO: Every page on the website is responsive except the Projects page. Projects is getting a complete redesign and will be ready to be viewed on mobile devices.