Added dynamic nested routes to Dashboard
Previously, there I couldn't figure out dynamic nested routing, so I made all of the dashboard components swap on click, but everything stays on the path /dashboard. This is not good when the user reloads because then the dashboard resets to the default component rather than the current component. It also isn't very good for SEO because if a crawler somehow is able to reach to /dashboard, it will only see the default path and no other paths. The second reason isn't too important because it is behind authentication so the crawler will probably not get past it, and it is only for viewing user data which shouldn't really be public. I figured out dynamic nested routes, which makes the user experience better. It is also a lot easier to read for developers because the code is a lot shorter and not as messy.