Add image upload to create event form and finished the event card list by connecting it with storage
Instead of creating the same downloadURL each time someone accesses the website, it is simpler to create the downloadURL and store it in the event document. It is also easier than dispatching an action to get each image, modifying the state with the image url, and mapping the new state to the props to render the image. That is a lot more complex and this method saves unnecessary calls.