Add attendance chart to dashboard
The attendance chart is an svg chart consisting of two lines, events and meetings. The user is able to view the entire dataset or select a region of the data by either zooming and panning on the main chart or selecting a region and "brushing" on the mini chart. Hovering over or nearby a point will display the event title, date, and number of people that attended/volunteered.

TODO: The next chart will be a single line chart of the change in active membership as time moves on. Since it will be another line chart with the same features, it may make more sense to create a line chart component.